requirements.txt
torch>=1.10.0
dgl>=0.8.0
rdkit-pypi>=2022.3.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
pillow>=9.0.0
pyyaml>=6.0
jarvis-tools>=2021.0.0

